chore: update dependencies and refactor navigation components
- Updated package versions in pnpm-lock.yaml and package.json for next, react, and react-dom.
- Refactored SiteNavBar component to utilize a centralized navLinks constant for better maintainability.
- Removed unused navigation links and simplified layout configuration.
- Added a new layout file for the playground section.
