+++ apps/website/package.json
@@ -27,9 +27,9 @@
         "fumadocs-ui": "^15.5.1",
         "lodash-es": "^4.17.21",
         "markdown-table": "^3.0.4",
-        "next": "^15.3.1",
-        "react": "^18.3.1",
-        "react-dom": "^18",
+        "next": "^15.3.4",
+        "react": "^18.0.0",
+        "react-dom": "^18.0.0",
         "remark": "^15.0.1",
         "remark-gfm": "^4.0.0",
         "remark-mdx": "^3.1.0",
@@ -44,8 +44,8 @@
         "@types/lodash-es": "^4.17.12",
         "@types/mdx": "^2.0.13",
         "@types/node": "^22.15.3",
-        "@types/react": "^18.3.18",
-        "@types/react-dom": "^18.3.5",
+        "@types/react": "^18.0.0",
+        "@types/react-dom": "^18.0.0",
         "eslint": "^9.20.0",
         "postcss": "^8.5.3",
         "shiki": "^3.6.0",
